[FLINK-13931][hive] Support Hive version 2.0.x
This closes #9644.
