+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/catalog/hive/client/HiveShimLoader.java
@@ -35,6 +35,8 @@ public class HiveShimLoader {
 	public static final String HIVE_VERSION_V1_2_0 = "1.2.0";
 	public static final String HIVE_VERSION_V1_2_1 = "1.2.1";
 	public static final String HIVE_VERSION_V1_2_2 = "1.2.2";
+	public static final String HIVE_VERSION_V2_0_0 = "2.0.0";
+	public static final String HIVE_VERSION_V2_0_1 = "2.0.1";
 	public static final String HIVE_VERSION_V2_1_0 = "2.1.0";
 	public static final String HIVE_VERSION_V2_1_1 = "2.1.1";
 	public static final String HIVE_VERSION_V2_2_0 = "2.2.0";
@@ -66,6 +68,12 @@ public class HiveShimLoader {
 			}
 			if (v.startsWith(HIVE_VERSION_V1_2_2)) {
 				return new HiveShimV122();
+			}
+			if (v.startsWith(HIVE_VERSION_V2_0_0)) {
+				return new HiveShimV200();
+			}
+			if (v.startsWith(HIVE_VERSION_V2_0_1)) {
+				return new HiveShimV201();
 			}
 			if (v.startsWith(HIVE_VERSION_V2_1_0)) {
 				return new HiveShimV210();
