+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/connectors/hive/HiveRunnerShimLoader.java
@@ -41,6 +41,8 @@ public class HiveRunnerShimLoader {
 				case HiveShimLoader.HIVE_VERSION_V1_2_1:
 				case HiveShimLoader.HIVE_VERSION_V1_2_2:
 					return new HiveRunnerShimV3();
+				case HiveShimLoader.HIVE_VERSION_V2_0_0:
+				case HiveShimLoader.HIVE_VERSION_V2_0_1:
 				case HiveShimLoader.HIVE_VERSION_V2_1_0:
 				case HiveShimLoader.HIVE_VERSION_V2_1_1:
 				case HiveShimLoader.HIVE_VERSION_V2_2_0:
